Allow to configure timeout duration by a provider
Hi, 

I really enjoy the work done on postman, it's beautiful and simple. 

I needed to change the duration of the timeout, so I propose to add a provider which can be use during the config phase (so it can be extended in the future for others parameters). 

If you agree with this, can you do the merge ?

Thank you